rename file purifyOptions to purify-options
