+++ packages/elements/purify-options.js
