+++ scripts/webpack/webpack-sass-prod/build-element-scss.js
@@ -2,7 +2,7 @@ const nodeSass = require('node-sass')
 const path = require('path')
 const glob = require('glob')
 const purifyCss = require('purify-css')
-const purifyOptions = require(path.resolve(__dirname, '../../../packages/elements/purifyOptions.js'))
+const purifyOptions = require(path.resolve(__dirname, '../../../packages/elements/purify-options.js'))
 
 /**
  * build css from scss imported from elements
